Moved date-time-converter module to remoting-common
