+++ platform-examples/lazy-loading-sample/lazy-loading-common/lazy-loading-common.gradle
@@ -16,5 +16,4 @@
 
 dependencies {
   compile project(':dolphin-platform-remoting-common')
-  compile project(':dolphin-platform-date-time-converter')
 }
