--- platform/dolphin-platform-date-time-converter/gradle.properties
@@ -1,1 +1,0 @@
-publishJars = true
