--- platform/dolphin-platform-date-time-converter/src/main/java/com/canoo/dolphin/converters/ValueFieldTypes.java
@@ -1,34 +1,0 @@
-/*
- * Copyright 2015-2017 Canoo Engineering AG.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.canoo.dolphin.converters;
-
-import org.apiguardian.api.API;
-
-import static org.apiguardian.api.API.Status.INTERNAL;
-
-@API(since = "0.x", status = INTERNAL)
-public interface ValueFieldTypes {
-
-    int DURATION_FIELD_TYPE = 51;
-
-    int LOCAL_DATE_TIME_FIELD_TYPE = 52;
-
-    int PERIODE_FIELD_TYPE = 53;
-
-    int ZONED_DATE_TIME_FIELD_TYPE = 54;
-    
-    int LOCAL_DATE_FIELD_TYPE = 55;
-}
