--- platform/dolphin-platform-date-time-converter/src/main/resources/META-INF/services/com.canoo.platform.remoting.spi.converter.ConverterFactory
@@ -1,5 +1,0 @@
-com.canoo.dolphin.converters.DurationConverterFactory
-com.canoo.dolphin.converters.LocalDateTimeConverterFactory
-com.canoo.dolphin.converters.PeriodConverterFactory
-com.canoo.dolphin.converters.ZonedDateTimeConverterFactory
-com.canoo.dolphin.converters.LocalDateConverterFactory
