+++ platform/dolphin-platform-remoting-common/src/test/java/com/canoo/dolphin/impl/converters/DurationConverterFactoryTest.java
@@ -13,22 +13,23 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.canoo.dolphin.converters;
+package com.canoo.dolphin.impl.converters;
 
 import com.canoo.platform.remoting.spi.converter.Converter;
 import com.canoo.platform.remoting.spi.converter.ValueConverterException;
 import com.canoo.dp.impl.remoting.Converters;
 import com.canoo.dp.impl.remoting.BeanRepository;
+import com.canoo.dp.impl.remoting.converters.ValueFieldTypes;
 import mockit.Mocked;
 import org.testng.annotations.Test;
 
-import java.time.Period;
+import java.time.Duration;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.fail;
 
-public class PeriodeConverterFactoryTest {
+public class DurationConverterFactoryTest {
 
     @Test
     public void testFactoryFieldType(@Mocked BeanRepository beanRepository) {
@@ -36,10 +37,10 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        int type = converters.getFieldType(Period.class);
+        int type = converters.getFieldType(Duration.class);
 
         //Then
-        assertEquals(type, ValueFieldTypes.PERIODE_FIELD_TYPE);
+        assertEquals(type, ValueFieldTypes.DURATION_FIELD_TYPE);
     }
 
     @Test
@@ -48,7 +49,7 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        Converter converter = converters.getConverter(Period.class);
+        Converter converter = converters.getConverter(Duration.class);
 
         //Then
         assertNotNull(converter);
@@ -60,14 +61,13 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        Converter converter = converters.getConverter(Period.class);
+        Converter converter = converters.getConverter(Duration.class);
 
         //Then
-        testReconversion(converter, Period.ofDays(7));
-        testReconversion(converter, Period.ofYears(700_000_000));
-        testReconversion(converter, Period.ofWeeks(70_000_000));
-        testReconversion(converter, Period.ZERO);
-        testReconversion(converter, Period.ofDays(10_000_000));
+        testReconversion(converter, Duration.ZERO);
+        testReconversion(converter, Duration.ofDays(7));
+        testReconversion(converter, Duration.ofMillis(10));
+        testReconversion(converter, Duration.ofDays(10_000_000));
     }
 
     @Test
@@ -76,14 +76,14 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        Converter converter = converters.getConverter(Period.class);
+        Converter converter = converters.getConverter(Duration.class);
 
         //Then
         try {
             assertEquals(converter.convertFromDolphin(null), null);
             assertEquals(converter.convertToDolphin(null), null);
         } catch (ValueConverterException e) {
-            fail("Error in conversion", e);
+            fail("Error in conversion");
         }
     }
 
@@ -93,7 +93,7 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        Converter converter = converters.getConverter(Period.class);
+        Converter converter = converters.getConverter(Duration.class);
 
         //Then
         converter.convertFromDolphin(7);
@@ -105,20 +105,20 @@ public class PeriodeConverterFactoryTest {
         Converters converters = new Converters(beanRepository);
 
         //When
-        Converter converter = converters.getConverter(Period.class);
+        Converter converter = converters.getConverter(Duration.class);
 
         //Then
         converter.convertToDolphin(7);
     }
 
-    private void testReconversion(Converter converter, Period duration) {
+    private void testReconversion(Converter converter, Duration duration) {
         try {
             Object dolphinObject = converter.convertToDolphin(duration);
             assertNotNull(dolphinObject);
             Object reconvertedObject = converter.convertFromDolphin(dolphinObject);
             assertNotNull(reconvertedObject);
-            assertEquals(reconvertedObject.getClass(), Period.class);
-            Period reconvertedDuration = (Period) reconvertedObject;
+            assertEquals(reconvertedObject.getClass(), Duration.class);
+            Duration reconvertedDuration = (Duration) reconvertedObject;
             assertEquals(reconvertedDuration, duration);
         } catch (ValueConverterException e) {
             fail("Error in conversion", e);
